refactor(rollback): split effectful shell into DI modules
Introduce a RollbackDeps seam (deps.ts) and split the imperative shell into
content-snapshot, page-mutation, activity-repo, preview, rollback-executors,
redo-executors, and execute modules — each taking { db, clock, genChangeGroupId,
readContent, syncMentions, createPageVersion, getActorInfo, logRollbackActivity,
canUserRollback, createMentionNotification, logger }. rollback-service.ts becomes
the composition root binding defaultRollbackDeps() and re-exporting the frozen
public API. options.tx becomes deps.db via withTx so one transaction threads
through every handler, createPageVersion and logRollbackActivity; the
revision-guarded write and error-swallowing reads are preserved; mention
notifications stay fire-and-forget. executeRollback now fetches the activity and
resolves its content snapshot exactly once and threads them through preview,
handler, and audit log. All rollback + ai-undo suites (285 tests) pass unchanged;
DI-seam tests cover the revision guard and error swallowing; typecheck green.
