+++ apps/web/src/services/api/rollback-service.ts
@@ -1,1066 +1,46 @@
 /**
- * Rollback Service
+ * Rollback Service — composition root.
  *
- * Handles version history rollback operations for PageSpace.
- * Allows users to restore resources to previous states based on activity logs.
+ * Handles version history rollback operations for PageSpace. The functional
+ * core (pure decision modules) and the imperative shell (effectful modules that
+ * take a RollbackDeps) live under services/api/rollback/. This file binds the
+ * shell to the production deps and re-exports the frozen public API so existing
+ * consumers (ai-undo-service, rollback-to-point-service) compile untouched.
  */
-
-import { db } from '@pagespace/db/db'
-import { eq, and, desc, gt, count, asc, not, inArray } from '@pagespace/db/operators'
-import { users } from '@pagespace/db/schema/auth'
-import { pages, drives } from '@pagespace/db/schema/core'
-import { activityLogs } from '@pagespace/db/schema/monitoring'
-import { driveMembers, driveRoles, pagePermissions } from '@pagespace/db/schema/members'
-import type { ActivityAction, ActivityActionPreview, ActivityActionResult } from '@/types/activity-actions';
+import type { db } from '@pagespace/db/db'
+import type { RollbackContext } from '@pagespace/lib/permissions/rollback-permissions'
+import type { ActivityActionPreview } from '@/types/activity-actions'
+import { defaultRollbackDeps } from './rollback/deps'
+import type { ActivityLogForRollback } from './rollback/types'
 import {
-  canUserRollback,
-  isRollbackableOperation,
-  type RollbackContext,
-} from '@pagespace/lib/permissions/rollback-permissions';
-
-// Re-export RollbackContext for consumers
-export type { RollbackContext };
-import { logRollbackActivity, getActorInfo, type ActivityResourceType, type ActivityOperation, type DeferredWorkflowTrigger } from '@pagespace/lib/monitoring/activity-logger'
-import { createChangeGroupId, inferChangeGroupType } from '@pagespace/lib/monitoring/change-group';
-import { loggers } from '@pagespace/lib/logging/logger-config'
-import { readPageContent } from '@pagespace/lib/services/page-content-store'
-import { createPageVersion, type PageVersionSource } from '@pagespace/lib/services/page-version-service'
-import type { ChangeGroupType } from '@pagespace/lib/monitoring/change-group';
-import type { PageContentFormat } from '@pagespace/lib/content/page-content-format';
-import { syncMentions, type SyncMentionsResult } from '@/services/api/page-mention-service';
-import { createMentionNotification } from '@pagespace/lib/notifications/notifications';
+  getActivityById as repoGetActivityById,
+  getPageVersionHistory as repoGetPageVersionHistory,
+  getDriveVersionHistory as repoGetDriveVersionHistory,
+  getUserRetentionDays as repoGetUserRetentionDays,
+  type VersionHistoryOptions,
+} from './rollback/activity-repo'
 import {
-  REDO_ALLOW_MISSING_TARGET,
-  ROLLBACK_ALLOW_MISSING_TARGET,
-  AGENT_CONFIG_ROLLBACK_FIELDS,
-} from './rollback/operations';
-import { getConflictFields, classifyUndoGroupConflict, isNoOpChange } from './rollback/conflict';
-import type { ActivityLogForRollback } from './rollback/types';
-import {
-  buildChangeSummary,
-  buildActionTargetValues,
-  getEffectiveOperation,
-  isRollingBackRollback,
-} from './rollback/target-values';
-import { mapActivityRow, buildHistoryConditions } from './rollback/activity-mapping';
-import { computePageMutation, pickConversationTable } from './rollback/page-mutation-plan';
-import { buildPreview, evaluateEligibility, evaluateCreateNoOp } from './rollback/preview-eligibility';
-import {
-  planPageRollback,
-  planDriveRollback,
-  planPermissionRollback,
-  planAgentRollback,
-  planMemberRollback,
-  planRoleRollback,
-  planMessageRollback,
-} from './rollback/rollback-plans';
-import {
-  planPageRedo,
-  planDriveRedo,
-  planPermissionRedo,
-  planAgentRedo,
-  planMemberRedo,
-  planRoleRedo,
-  planMessageRedo,
-} from './rollback/redo-plans';
+  executeRollback as shellExecuteRollback,
+  previewRollback as shellPreviewRollback,
+  type RollbackResult,
+} from './rollback/execute'
 
-// Re-export the shared activity shape so existing consumers keep importing it here.
-export type { ActivityLogForRollback };
+// Re-export the frozen public types.
+export type { RollbackContext }
+export type { ActivityLogForRollback }
+export type { VersionHistoryOptions }
+export type { RollbackResult }
+/** Result of a rollback preview */
+export type RollbackPreview = ActivityActionPreview
 
-async function resolveActivityContentSnapshot(activity: ActivityLogForRollback): Promise<string | null> {
-  if (activity.contentSnapshot) {
-    return activity.contentSnapshot;
-  }
-
-  if (activity.contentRef) {
-    try {
-      return await readPageContent(activity.contentRef);
-    } catch (error) {
-      loggers.api.warn('[RollbackService] Failed to read content snapshot', {
-        activityId: activity.id,
-        contentRef: activity.contentRef,
-        error: error instanceof Error ? error.message : String(error),
-      });
-      return null;
-    }
-  }
-
-  return null;
-}
-
-/**
- * Resolve the effective conflict fields for a preview: detect raw conflicts,
- * then (when an undo group is in play) query for a later modification made
- * *outside* the group and classify accordingly. The DB lookup is the only
- * effect here — the branch decision lives in the pure classifyUndoGroupConflict.
- */
-async function resolveConflictFields(
-  database: typeof db,
-  activity: ActivityLogForRollback,
-  currentValues: Record<string, unknown> | null,
-  resourceType: ActivityResourceType,
-  undoGroupActivityIds: string[]
-): Promise<string[]> {
-  const conflictFields = getConflictFields(activity.newValues, currentValues);
-  if (conflictFields.length === 0) {
-    return conflictFields;
-  }
-
-  loggers.api.debug('[Rollback:Preview] Conflict detected', {
-    activityId: activity.id,
-    resourceId: activity.resourceId,
-    resourceType,
-    timestamp: activity.timestamp?.toISOString(),
-    conflictFields,
-    undoGroupSize: undoGroupActivityIds.length,
-  });
-
-  if (undoGroupActivityIds.length === 0) {
-    return classifyUndoGroupConflict({
-      conflictFields,
-      hasUndoGroupContext: false,
-      hasExternalModification: false,
-    });
-  }
-
-  const externalModifications = await database
-    .select({ id: activityLogs.id })
-    .from(activityLogs)
-    .where(
-      and(
-        eq(activityLogs.resourceId, activity.resourceId),
-        eq(activityLogs.resourceType, resourceType),
-        gt(activityLogs.timestamp, activity.timestamp),
-        not(inArray(activityLogs.id, undoGroupActivityIds))
-      )
-    )
-    .limit(1);
-
-  const resolved = classifyUndoGroupConflict({
-    conflictFields,
-    hasUndoGroupContext: true,
-    hasExternalModification: externalModifications.length > 0,
-  });
-
-  if (resolved.length === 0) {
-    loggers.api.debug('[Rollback:Preview] Conflict is internal to undo group, ignoring', {
-      activityId: activity.id,
-      resourceType,
-      conflictFields,
-    });
-  }
-
-  return resolved;
-}
-
-interface PageUpdateWithRevisionOptions {
-  userId?: string | null;
-  changeGroupId?: string;
-  changeGroupType?: ChangeGroupType;
-  source?: PageVersionSource;
-  metadata?: Record<string, unknown>;
-}
-
-interface PageUpdateContext {
-  userId: string;
-  changeGroupId: string;
-  changeGroupType: ChangeGroupType;
-  source: PageVersionSource;
-  metadata?: Record<string, unknown>;
-}
-
-interface PageMutationMeta {
-  pageId: string;
-  nextRevision: number;
-  stateHashBefore: string;
-  stateHashAfter: string;
-  contentRefAfter: string | null;
-  contentSizeAfter: number | null;
-  contentFormatAfter: PageContentFormat;
-  mentionsResult?: SyncMentionsResult;
-}
-
-interface PageChangeResult {
-  restoredValues: Record<string, unknown>;
-  pageMutationMeta: PageMutationMeta | undefined;
-}
-
-async function applyPageUpdateWithRevision(
-  database: typeof db,
-  pageId: string,
-  updateData: Record<string, unknown>,
-  options?: PageUpdateWithRevisionOptions
-): Promise<PageMutationMeta> {
-  const [currentPage] = await database
-    .select()
-    .from(pages)
-    .where(eq(pages.id, pageId))
-    .limit(1);
-
-  if (!currentPage) {
-    throw new Error('Page not found');
-  }
-
-  const {
-    currentRevision,
-    nextRevision,
-    nextContent,
-    contentFormatAfter,
-    contentRefAfter,
-    stateHashBefore,
-    stateHashAfter,
-  } = computePageMutation(currentPage, updateData);
-
-  const [updated] = await database
-    .update(pages)
-    .set({
-      ...updateData,
-      revision: nextRevision,
-      stateHash: stateHashAfter,
-      updatedAt: new Date(),
-    })
-    .where(
-      and(
-        eq(pages.id, pageId),
-        eq(pages.revision, currentRevision)
-      )
-    )
-    .returning({ id: pages.id });
-
-  if (!updated) {
-    throw new Error('Page was modified while applying rollback');
-  }
-
-  let mentionsResult: SyncMentionsResult | undefined;
-  if (updateData.content !== undefined) {
-    mentionsResult = await syncMentions(pageId, nextContent, database, { mentionedByUserId: options?.userId ?? undefined });
-  }
-
-  const changeGroupId = options?.changeGroupId ?? createChangeGroupId();
-  const changeGroupType = options?.changeGroupType ?? inferChangeGroupType({ isAiGenerated: false });
-
-  const version = await createPageVersion({
-    pageId,
-    driveId: currentPage.driveId,
-    createdBy: options?.userId ?? null,
-    source: options?.source ?? 'restore',
-    content: nextContent,
-    contentFormat: contentFormatAfter,
-    pageRevision: nextRevision,
-    stateHash: stateHashAfter,
-    changeGroupId,
-    changeGroupType,
-    metadata: options?.metadata,
-  }, { tx: database });
-
-  return {
-    pageId,
-    nextRevision,
-    stateHashBefore,
-    stateHashAfter,
-    contentRefAfter: version.contentRef ?? contentRefAfter ?? null,
-    contentSizeAfter: version.contentSize ?? null,
-    contentFormatAfter,
-    mentionsResult,
-  };
-}
-
-/**
- * Result of a rollback preview
- */
-export type RollbackPreview = ActivityActionPreview;
-
-/**
- * Result of executing a rollback
- */
-export interface RollbackResult extends ActivityActionResult {
-  success: boolean;
-  rollbackActivityId?: string;
-  restoredValues?: Record<string, unknown>;
-  /** Deferred workflow trigger to fire after transaction commit */
-  deferredWorkflowTrigger?: DeferredWorkflowTrigger;
-}
-
-/**
- * Options for fetching version history
- */
-export interface VersionHistoryOptions {
-  limit?: number;
-  offset?: number;
-  startDate?: Date;
-  endDate?: Date;
-  actorId?: string;
-  operation?: string;
-  includeAiOnly?: boolean;
-  resourceType?: string;
-}
+// Production dependency bag (real db, clock, and effect functions).
+const deps = defaultRollbackDeps()
 
 /**
  * Fetch a single activity log by ID
  */
-export async function getActivityById(
-  activityId: string
-): Promise<ActivityLogForRollback | null> {
-  loggers.api.debug('[Rollback:Fetch] Fetching activity by ID', { activityId });
-
-  try {
-    const result = await db
-      .select()
-      .from(activityLogs)
-      .where(eq(activityLogs.id, activityId))
-      .limit(1);
-
-    if (result.length === 0) {
-      loggers.api.debug('[Rollback:Fetch] Activity not found', { activityId });
-      return null;
-    }
-
-    const activity = result[0];
-    loggers.api.debug('[Rollback:Fetch] Activity found', {
-      activityId,
-      operation: activity.operation,
-      resourceType: activity.resourceType,
-      resourceId: activity.resourceId,
-      isAiGenerated: activity.isAiGenerated,
-    });
-
-    return mapActivityRow(activity);
-  } catch (error) {
-    loggers.api.error('[RollbackService] Error fetching activity', {
-      activityId,
-      error: error instanceof Error ? error.message : String(error),
-    });
-    return null;
-  }
-}
-
-/**
- * Preview what a rollback or redo would do
- */
-async function previewActivityAction(
-  action: ActivityAction,
-  activityId: string,
-  userId: string,
-  context: RollbackContext,
-  options?: { force?: boolean; undoGroupActivityIds?: string[] }
-): Promise<ActivityActionPreview> {
-  const force = options?.force ?? false;
-  const undoGroupActivityIds = options?.undoGroupActivityIds ?? [];
-  loggers.api.debug('[Rollback:Preview] Starting preview', { action, activityId, userId, context, force });
-
-  const activity = await getActivityById(activityId);
-  const resolvedContentSnapshot = activity
-    ? await resolveActivityContentSnapshot(activity)
-    : null;
-  const targetValues = activity ? buildActionTargetValues(activity, resolvedContentSnapshot) : null;
-  const changes = activity ? buildChangeSummary(activity, targetValues) : [];
-  const affectedResources = activity
-    ? [
-        {
-          type: activity.resourceType,
-          id: activity.resourceId,
-          title: activity.resourceTitle || 'Untitled',
-        },
-      ]
-    : [];
-
-  const basePreview = (overrides: Partial<ActivityActionPreview>): ActivityActionPreview =>
-    buildPreview({ action, targetValues, changes, affectedResources }, overrides);
-
-  if (!activity) {
-    loggers.api.debug('[Rollback:Preview] Activity not found', { activityId });
-    return basePreview({ reason: 'Activity not found' });
-  }
-
-  // When rolling back a rollback activity, we need the source operation to know what handler to use
-  const rollingBackRollback = isRollingBackRollback(activity);
-
-  const effectiveOperation = getEffectiveOperation(activity);
-  const isRollbackable = effectiveOperation ? isRollbackableOperation(effectiveOperation) : false;
-  const hasTargetValues = !!targetValues && Object.keys(targetValues).length > 0;
-  // Content snapshots are only relevant for regular rollbacks, not rollback of rollbacks
-  const hasContentSnapshot = !rollingBackRollback && !!resolvedContentSnapshot;
-  // When rolling back a rollback, use the redo allow list since we're restoring forward
-  const allowMissingTarget = !!effectiveOperation && (rollingBackRollback
-    ? REDO_ALLOW_MISSING_TARGET.has(effectiveOperation)
-    : ROLLBACK_ALLOW_MISSING_TARGET.has(effectiveOperation));
-  loggers.api.debug('[Rollback:Preview] Checking eligibility', {
-    action,
-    operation: effectiveOperation,
-    isRollbackable,
-    rollingBackRollback,
-    hasTargetValues,
-    hasContentSnapshot,
-    allowMissingTarget,
-    previousValuesFields: targetValues ? Object.keys(targetValues) : [],
-  });
-
-  const eligibility = evaluateEligibility({
-    action,
-    rollingBackRollback,
-    effectiveOperation,
-    isRollbackable,
-    hasTargetValues,
-    hasContentSnapshot,
-    allowMissingTarget,
-  });
-  if (eligibility.kind === 'reject') {
-    return basePreview({ reason: eligibility.reason });
-  }
-
-  // Check permissions
-  loggers.api.debug('[Rollback:Preview] Checking permissions', {
-    userId,
-    context,
-    resourceType: activity.resourceType,
-  });
-
-  const permissionCheck = await canUserRollback(userId, activity, context);
-
-  loggers.api.debug('[Rollback:Preview] Permission check result', {
-    canRollback: permissionCheck.canRollback,
-    reason: permissionCheck.reason,
-  });
-
-  if (!permissionCheck.canRollback) {
-    return basePreview({
-      reason: permissionCheck.reason,
-    });
-  }
-
-  // Get current state and check for conflicts
-  const warnings: string[] = [];
-  let currentValues: Record<string, unknown> | null = null;
-  let conflictFields: string[] = [];
-  let hasConflict = false;
-  let requiresForce = false;
-
-  loggers.api.debug('[Rollback:Preview] Fetching current resource state', {
-    resourceType: activity.resourceType,
-    resourceId: activity.resourceId,
-  });
-
-  if (activity.resourceType === 'page' && activity.pageId) {
-    const currentPage = await db
-      .select()
-      .from(pages)
-      .where(eq(pages.id, activity.pageId))
-      .limit(1);
-
-    if (currentPage.length === 0) {
-      return basePreview({
-        reason: 'Resource no longer exists',
-      });
-    }
-
-    // Check if parent drive still exists and is not trashed
-    if (activity.driveId) {
-      const parentDrive = await db
-        .select({ id: drives.id, isTrashed: drives.isTrashed })
-        .from(drives)
-        .where(eq(drives.id, activity.driveId))
-        .limit(1);
-
-      if (parentDrive.length === 0) {
-        return basePreview({
-          reason: 'Parent drive has been deleted',
-        });
-      }
-
-      if (parentDrive[0].isTrashed) {
-        return basePreview({
-          reason: 'Parent drive is in trash. Restore the drive first.',
-        });
-      }
-    }
-
-    currentValues = {
-      title: currentPage[0].title,
-      content: currentPage[0].content,
-      parentId: currentPage[0].parentId,
-      position: currentPage[0].position,
-      isTrashed: currentPage[0].isTrashed,
-    };
-
-    // Check if create operation is a no-op (page already in desired state)
-    if (effectiveOperation === 'create') {
-      // When rolling back a rollback of create, we should RESTORE (not trash)
-      const shouldBeTrashed = action === 'rollback' && !rollingBackRollback;
-      const createNoOp = evaluateCreateNoOp({
-        isCreateNoOp: currentPage[0].isTrashed === shouldBeTrashed,
-        shouldBeTrashed,
-        // For AI undo, a no-op is silently skippable so stale activities from
-        // previous sessions can be filtered out; outside an undo group it errors.
-        hasUndoGroup: undoGroupActivityIds.length > 0,
-        allowUndoGroupSkip: true,
-        trashedReason: 'Page is already in trash',
-        restoredReason: 'Page is already restored',
-      });
-      if (createNoOp.kind === 'skippable') {
-        return basePreview({ canExecute: true, isNoOp: true, currentValues });
-      }
-      if (createNoOp.kind === 'error') {
-        return basePreview({ reason: createNoOp.reason, currentValues, isNoOp: true });
-      }
-    }
-
-    // For 'create' operations, skip conflict check - we're just trashing the page
-    // For update operations, check for conflicts and distinguish internal vs external
-    if (effectiveOperation !== 'create') {
-      conflictFields = await resolveConflictFields(db, activity, currentValues, 'page', undoGroupActivityIds);
-    }
-
-    if (conflictFields.length > 0) {
-      hasConflict = true;
-      requiresForce = true;
-      loggers.api.debug('[Rollback:Preview] Conflict check', {
-        hasConflict: true,
-        checkedFields: conflictFields,
-        force,
-      });
-
-      if (!force) {
-        return basePreview({
-          reason: 'Resource has been modified since this change. Use force=true to override.',
-          currentValues,
-          hasConflict,
-          conflictFields,
-          requiresForce,
-        });
-      }
-      warnings.push('This resource has been modified since this change. Recent changes will be overwritten.');
-    }
-  } else if (activity.resourceType === 'drive' && activity.driveId) {
-    const currentDrive = await db
-      .select()
-      .from(drives)
-      .where(eq(drives.id, activity.driveId))
-      .limit(1);
-
-    if (currentDrive.length === 0) {
-      return basePreview({
-        reason: 'Drive no longer exists',
-      });
-    }
-
-    currentValues = {
-      name: currentDrive[0].name,
-      isTrashed: currentDrive[0].isTrashed,
-      drivePrompt: currentDrive[0].drivePrompt,
-      ownerId: currentDrive[0].ownerId,
-    };
-
-    if (effectiveOperation === 'create') {
-      // When rolling back a rollback of create, we should RESTORE (not trash)
-      const shouldBeTrashed = action === 'rollback' && !rollingBackRollback;
-      const createNoOp = evaluateCreateNoOp({
-        isCreateNoOp: currentDrive[0].isTrashed === shouldBeTrashed,
-        shouldBeTrashed,
-        // Drive create no-ops always error (no silent undo-group skip).
-        hasUndoGroup: undoGroupActivityIds.length > 0,
-        allowUndoGroupSkip: false,
-        trashedReason: 'Drive is already in trash',
-        restoredReason: 'Drive is already restored',
-      });
-      if (createNoOp.kind === 'error') {
-        return basePreview({
-          reason: createNoOp.reason,
-          currentValues,
-          isNoOp: true,
-        });
-      }
-    }
-
-    // Check for conflicts and distinguish internal vs external
-    conflictFields = await resolveConflictFields(db, activity, currentValues, 'drive', undoGroupActivityIds);
-
-    if (conflictFields.length > 0) {
-      hasConflict = true;
-      requiresForce = true;
-      loggers.api.debug('[Rollback:Preview] Drive conflict check', {
-        hasConflict: true,
-        checkedFields: conflictFields,
-        force,
-      });
-
-      if (!force) {
-        return basePreview({
-          reason: 'Drive has been modified since this change. Use force=true to override.',
-          currentValues,
-          hasConflict,
-          conflictFields,
-          requiresForce,
-        });
-      }
-      warnings.push('This drive has been modified since this change. Recent changes will be overwritten.');
-    }
-  } else if (activity.resourceType === 'member' && activity.driveId) {
-    const metadata = activity.metadata as { targetUserId?: string } | null;
-    const targetUserId = metadata?.targetUserId || (activity.previousValues?.userId as string);
-
-    if (targetUserId) {
-      const currentMember = await db
-        .select()
-        .from(driveMembers)
-        .where(and(
-          eq(driveMembers.driveId, activity.driveId),
-          eq(driveMembers.userId, targetUserId)
-        ))
-        .limit(1);
-
-      if (currentMember.length > 0) {
-        currentValues = {
-          userId: targetUserId,
-          role: currentMember[0].role,
-          customRoleId: currentMember[0].customRoleId,
-          invitedBy: currentMember[0].invitedBy,
-          invitedAt: currentMember[0].invitedAt,
-          acceptedAt: currentMember[0].acceptedAt,
-        };
-      }
-
-      if (effectiveOperation === 'member_add') {
-        if (rollingBackRollback) {
-          // Rollback of rollback: re-adding member, no-op if already exists
-          if (currentMember.length > 0) {
-            return basePreview({
-              reason: 'Member is already in the drive',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        } else {
-          // Regular rollback: removing member, no-op if already removed
-          if (currentMember.length === 0) {
-            return basePreview({
-              reason: 'Member has already been removed',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        }
-      }
-
-      if (effectiveOperation === 'member_remove') {
-        if (rollingBackRollback) {
-          // Rollback of rollback: removing member, no-op if already removed
-          if (currentMember.length === 0) {
-            return basePreview({
-              reason: 'Member has already been removed',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        } else {
-          // Regular rollback: re-adding member, no-op if already exists
-          if (currentMember.length > 0) {
-            return basePreview({
-              reason: 'Member has already been re-added to the drive',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        }
-      }
-
-      if (effectiveOperation === 'member_role_change' && currentMember.length === 0) {
-        return basePreview({
-          reason: 'Member no longer exists',
-          currentValues,
-        });
-      }
-
-      if (currentMember.length > 0 && effectiveOperation === 'member_role_change') {
-        // Check for conflicts and distinguish internal vs external
-        conflictFields = await resolveConflictFields(db, activity, currentValues, 'member', undoGroupActivityIds);
-
-        if (conflictFields.length > 0) {
-          hasConflict = true;
-          requiresForce = true;
-          loggers.api.debug('[Rollback:Preview] Member conflict check', {
-            hasConflict: true,
-            force,
-          });
-
-          if (!force) {
-            return basePreview({
-              reason: 'Member role has been changed since this update. Use force=true to override.',
-              currentValues,
-              hasConflict,
-              conflictFields,
-              requiresForce,
-            });
-          }
-          warnings.push("This member's role has been changed since this update. Recent changes will be overwritten.");
-        }
-      }
-    }
-  } else if (activity.resourceType === 'permission' && activity.pageId) {
-    const metadata = activity.metadata as { targetUserId?: string } | null;
-    const targetUserId = metadata?.targetUserId || (activity.previousValues?.userId as string);
-
-    if (targetUserId) {
-      const currentPermission = await db
-        .select()
-        .from(pagePermissions)
-        .where(and(
-          eq(pagePermissions.pageId, activity.pageId),
-          eq(pagePermissions.userId, targetUserId)
-        ))
-        .limit(1);
-
-      if (currentPermission.length > 0) {
-        currentValues = {
-          userId: targetUserId,
-          canView: currentPermission[0].canView,
-          canEdit: currentPermission[0].canEdit,
-          canShare: currentPermission[0].canShare,
-          canDelete: currentPermission[0].canDelete,
-          note: currentPermission[0].note,
-          expiresAt: currentPermission[0].expiresAt,
-          grantedBy: currentPermission[0].grantedBy,
-        };
-      }
-
-      if (effectiveOperation === 'permission_grant') {
-        if (rollingBackRollback) {
-          // Rollback of rollback: re-granting permission, no-op if already exists
-          if (currentPermission.length > 0) {
-            return basePreview({
-              reason: 'Permission is already granted',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        } else {
-          // Regular rollback: revoking permission, no-op if already revoked
-          if (currentPermission.length === 0) {
-            return basePreview({
-              reason: 'Permission has already been revoked',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        }
-      }
-
-      if (effectiveOperation === 'permission_revoke') {
-        if (rollingBackRollback) {
-          // Rollback of rollback: revoking permission, no-op if already revoked
-          if (currentPermission.length === 0) {
-            return basePreview({
-              reason: 'Permission has already been revoked',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        } else {
-          // Regular rollback: re-granting permission, no-op if already exists
-          if (currentPermission.length > 0) {
-            return basePreview({
-              reason: 'Permission has already been re-granted',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        }
-      }
-
-      if (effectiveOperation === 'permission_update' && currentPermission.length === 0) {
-        return basePreview({
-          reason: 'Permission no longer exists',
-          currentValues,
-        });
-      }
-
-      if (currentPermission.length > 0 && effectiveOperation === 'permission_update') {
-        // Check for conflicts and distinguish internal vs external
-        conflictFields = await resolveConflictFields(db, activity, currentValues, 'permission', undoGroupActivityIds);
-
-        if (conflictFields.length > 0) {
-          hasConflict = true;
-          requiresForce = true;
-          loggers.api.debug('[Rollback:Preview] Permission conflict check', {
-            hasConflict: true,
-            force,
-          });
-
-          if (!force) {
-            return basePreview({
-              reason: 'Permissions have been changed since this update. Use force=true to override.',
-              currentValues,
-              hasConflict,
-              conflictFields,
-              requiresForce,
-            });
-          }
-          warnings.push('These permissions have been changed since this update. Recent changes will be overwritten.');
-        }
-      }
-    }
-  } else if (activity.resourceType === 'role' && activity.driveId) {
-    const metadata = activity.metadata as { roleId?: string } | null;
-    const roleId = metadata?.roleId || activity.resourceId;
-
-    // Role reorder affects all roles in the drive
-    if (effectiveOperation === 'role_reorder') {
-      // Get current order of roles in this drive
-      const currentRoles = await db
-        .select({ id: driveRoles.id, position: driveRoles.position })
-        .from(driveRoles)
-        .where(eq(driveRoles.driveId, activity.driveId))
-        .orderBy(asc(driveRoles.position));
-
-      currentValues = {
-        order: currentRoles.map(role => role.id),
-      };
-
-      // Skip conflict detection for AI undo - all changes are from the same conversation
-      if (undoGroupActivityIds.length === 0) {
-        conflictFields = getConflictFields(activity.newValues, currentValues);
-      }
-      if (conflictFields.length > 0) {
-        hasConflict = true;
-        requiresForce = true;
-        loggers.api.debug('[Rollback:Preview] Role reorder conflict check', {
-          hasConflict: true,
-          force,
-        });
-
-        if (!force) {
-          return basePreview({
-            reason: 'Roles have been reordered since this change. Use force=true to override.',
-            currentValues,
-            hasConflict,
-            conflictFields,
-            requiresForce,
-          });
-        }
-        warnings.push('Roles have been reordered since this change. Recent changes will be overwritten.');
-      }
-    } else if (roleId) {
-      const currentRole = await db
-        .select()
-        .from(driveRoles)
-        .where(eq(driveRoles.id, roleId))
-        .limit(1);
-
-      if (currentRole.length > 0) {
-        currentValues = {
-          name: currentRole[0].name,
-          description: currentRole[0].description,
-          color: currentRole[0].color,
-          isDefault: currentRole[0].isDefault,
-          permissions: currentRole[0].permissions,
-          position: currentRole[0].position,
-        };
-      }
-
-      if (effectiveOperation === 'create') {
-        if (rollingBackRollback) {
-          // Rollback of rollback: re-creating role, no-op if already exists
-          if (currentRole.length > 0) {
-            return basePreview({
-              reason: 'Role already exists',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        } else {
-          // Regular rollback: deleting role, no-op if already deleted
-          if (currentRole.length === 0) {
-            return basePreview({
-              reason: 'Role has already been deleted',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        }
-      }
-
-      if (effectiveOperation === 'delete') {
-        if (rollingBackRollback) {
-          // Rollback of rollback: deleting role, no-op if already deleted
-          if (currentRole.length === 0) {
-            return basePreview({
-              reason: 'Role has already been deleted',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        } else {
-          // Regular rollback: re-creating role, no-op if already exists
-          if (currentRole.length > 0) {
-            return basePreview({
-              reason: 'Role already exists with this ID',
-              currentValues,
-              isNoOp: true,
-            });
-          }
-        }
-      }
-
-      if (currentRole.length > 0 && effectiveOperation === 'update') {
-        // Check for conflicts and distinguish internal vs external
-        conflictFields = await resolveConflictFields(db, activity, currentValues, 'role', undoGroupActivityIds);
-
-        if (conflictFields.length > 0) {
-          hasConflict = true;
-          requiresForce = true;
-          loggers.api.debug('[Rollback:Preview] Role conflict check', {
-            hasConflict: true,
-            force,
-          });
-
-          if (!force) {
-            return basePreview({
-              reason: 'Role has been modified since this update. Use force=true to override.',
-              currentValues,
-              hasConflict,
-              conflictFields,
-              requiresForce,
-            });
-          }
-          warnings.push('This role has been modified since this update. Recent changes will be overwritten.');
-        }
-      }
-    }
-  } else if (activity.resourceType === 'agent' && activity.pageId) {
-    const currentAgent = await db
-      .select({
-        systemPrompt: pages.systemPrompt,
-        enabledTools: pages.enabledTools,
-        aiProvider: pages.aiProvider,
-        aiModel: pages.aiModel,
-        includeDrivePrompt: pages.includeDrivePrompt,
-        agentDefinition: pages.agentDefinition,
-        visibleToGlobalAssistant: pages.visibleToGlobalAssistant,
-        includePageTree: pages.includePageTree,
-        pageTreeScope: pages.pageTreeScope,
-        toolExposureMode: pages.toolExposureMode,
-        userScopedAccess: pages.userScopedAccess,
-      })
-      .from(pages)
-      .where(eq(pages.id, activity.pageId))
-      .limit(1);
-
-    if (currentAgent.length === 0) {
-      return basePreview({
-        reason: 'Agent no longer exists',
-      });
-    }
-
-    currentValues = {
-      systemPrompt: currentAgent[0].systemPrompt,
-      enabledTools: currentAgent[0].enabledTools,
-      aiProvider: currentAgent[0].aiProvider,
-      aiModel: currentAgent[0].aiModel,
-      includeDrivePrompt: currentAgent[0].includeDrivePrompt,
-      agentDefinition: currentAgent[0].agentDefinition,
-      visibleToGlobalAssistant: currentAgent[0].visibleToGlobalAssistant,
-      includePageTree: currentAgent[0].includePageTree,
-      pageTreeScope: currentAgent[0].pageTreeScope,
-      toolExposureMode: currentAgent[0].toolExposureMode,
-      userScopedAccess: currentAgent[0].userScopedAccess,
-    };
-
-    // Check for conflicts and distinguish internal vs external
-    conflictFields = await resolveConflictFields(db, activity, currentValues, 'agent', undoGroupActivityIds);
-
-    if (conflictFields.length > 0) {
-      hasConflict = true;
-      requiresForce = true;
-      loggers.api.debug('[Rollback:Preview] Agent conflict check', {
-        hasConflict: true,
-        force,
-      });
-
-      if (!force) {
-        return basePreview({
-          reason: 'Agent settings have been modified since this update. Use force=true to override.',
-          currentValues,
-          hasConflict,
-          conflictFields,
-          requiresForce,
-        });
-      }
-      warnings.push('Agent settings have been modified since this update. Recent changes will be overwritten.');
-    }
-  } else if (activity.resourceType === 'message') {
-    const metadata = activity.metadata as Record<string, unknown> | null;
-    const conversationType = metadata?.conversationType as string | undefined;
-    const { table } = pickConversationTable({ conversationType, hasPageId: !!activity.pageId });
-
-    const currentMessage = await db
-      .select()
-      .from(table)
-      .where(eq(table.id, activity.resourceId))
-      .limit(1);
-
-    if (currentMessage.length === 0) {
-      return basePreview({
-        reason: 'Message no longer exists',
-      });
-    }
-
-    currentValues = {
-      content: currentMessage[0].content,
-      isActive: currentMessage[0].isActive,
-    };
-
-    // Check for conflicts and distinguish internal vs external
-    conflictFields = await resolveConflictFields(db, activity, currentValues, 'message', undoGroupActivityIds);
-
-    if (conflictFields.length > 0) {
-      hasConflict = true;
-      requiresForce = true;
-      loggers.api.debug('[Rollback:Preview] Message conflict check', {
-        hasConflict: true,
-        force,
-      });
-
-      if (!force) {
-        return basePreview({
-          reason: 'Message has been modified since this change. Use force=true to override.',
-          currentValues,
-          hasConflict,
-          conflictFields,
-          requiresForce,
-        });
-      }
-      warnings.push('This message has been modified since this change. Recent changes will be overwritten.');
-    }
-  }
-
-  // Skip no-op detection for AI undo - operations may appear as no-ops
-  // but the resource will be affected by other operations in the undo group
-  const noOp = undoGroupActivityIds.length === 0 && isNoOpChange(targetValues, currentValues);
-  if (noOp) {
-    return basePreview({
-      reason: 'Already at this version',
-      currentValues,
-      warnings,
-      hasConflict,
-      conflictFields,
-      requiresForce,
-      isNoOp: true,
-    });
-  }
-
-  loggers.api.debug('[Rollback:Preview] Preview complete', {
-    canExecute: true,
-    warningsCount: warnings.length,
-    targetFieldsCount: targetValues ? Object.keys(targetValues).length : 0,
-  });
-
-  return basePreview({
-    canExecute: true,
-    currentValues,
-    warnings,
-    hasConflict,
-    conflictFields,
-    requiresForce,
-  });
+export async function getActivityById(activityId: string): Promise<ActivityLogForRollback | null> {
+  return repoGetActivityById(deps, activityId)
 }
 
 /**
@@ -1072,7 +52,7 @@ export async function previewRollback(
   context: RollbackContext,
   options?: { force?: boolean; undoGroupActivityIds?: string[] }
 ): Promise<RollbackPreview> {
-  return previewActivityAction('rollback', activityId, userId, context, options);
+  return shellPreviewRollback(deps, activityId, userId, context, options)
 }
 
 /**
@@ -1086,970 +66,7 @@ export async function executeRollback(
   context: RollbackContext,
   options?: { tx?: typeof db; force?: boolean; undoGroupActivityIds?: string[] }
 ): Promise<RollbackResult> {
-  const { tx, force, undoGroupActivityIds } = options ?? {};
-  loggers.api.debug('[Rollback:Execute] Starting execution', {
-    activityId,
-    userId,
-    context,
-    usingTransaction: !!tx,
-    force,
-  });
-
-  const preview = await previewRollback(activityId, userId, context, { force, undoGroupActivityIds });
-
-  if (!preview.canExecute) {
-    loggers.api.debug('[Rollback:Execute] Aborting - preview check failed', {
-      canExecute: preview.canExecute,
-      reason: preview.reason,
-      isNoOp: preview.isNoOp,
-    });
-    return {
-      success: preview.isNoOp,
-      action: 'rollback',
-      status: preview.isNoOp ? 'no_op' : 'failed',
-      message: preview.reason || 'Cannot rollback this activity',
-      warnings: preview.warnings,
-      changesApplied: preview.changes,
-    };
-  }
-
-  const activity = await getActivityById(activityId);
-  if (!activity) {
-    return {
-      success: false,
-      action: 'rollback',
-      status: 'failed',
-      message: 'Activity not found',
-      warnings: preview.warnings,
-      changesApplied: preview.changes,
-    };
-  }
-
-  const warnings: string[] = [...preview.warnings];
-  const database = tx ?? db;
-  const changeGroupId = createChangeGroupId();
-  const changeGroupType = inferChangeGroupType({ isAiGenerated: false });
-
-  // When rolling back a rollback, use the original source operation for metadata
-  const rollingBackRollback = isRollingBackRollback(activity);
-  if (rollingBackRollback && !activity.rollbackSourceOperation) {
-    return {
-      success: false,
-      action: 'rollback',
-      status: 'failed',
-      message: 'Rollback source operation not available for this activity',
-      warnings: preview.warnings,
-      changesApplied: preview.changes,
-    };
-  }
-  const effectiveSourceOperation = rollingBackRollback
-    ? activity.rollbackSourceOperation as ActivityOperation
-    : activity.operation as ActivityOperation;
-
-  const pageUpdateContext: PageUpdateContext = {
-    userId,
-    changeGroupId,
-    changeGroupType,
-    source: 'restore',
-    metadata: {
-      action: 'rollback',
-      rollbackFromActivityId: activityId,
-      rollbackSourceOperation: effectiveSourceOperation,
-      rollbackSourceTimestamp: activity.timestamp,
-    },
-  };
-
-  try {
-    // Get actor info for logging
-    const actorInfo = await getActorInfo(userId);
-
-    // Execute rollback based on resource type
-    let restoredValues: Record<string, unknown> = {};
-    let pageMutationMeta: PageMutationMeta | undefined;
-
-    loggers.api.debug('[Rollback:Execute] Executing handler', {
-      resourceType: activity.resourceType,
-      operation: activity.operation,
-      effectiveSourceOperation,
-      rollingBackRollback,
-      resourceId: activity.resourceId,
-    });
-
-    // When rolling back a rollback, use redo handlers (they take explicit targetValues and sourceOperation)
-    // Otherwise use regular rollback handlers
-    switch (activity.resourceType) {
-      case 'page': {
-        const result = rollingBackRollback
-          ? await redoPageChange(activity, preview.targetValues, effectiveSourceOperation, database, pageUpdateContext)
-          : await rollbackPageChange(activity, preview.currentValues, database, pageUpdateContext);
-        restoredValues = result.restoredValues;
-        pageMutationMeta = result.pageMutationMeta;
-        break;
-      }
-
-      case 'drive':
-        restoredValues = rollingBackRollback
-          ? await redoDriveChange(activity, preview.targetValues, effectiveSourceOperation, database, pageUpdateContext)
-          : await rollbackDriveChange(activity, preview.currentValues, database, pageUpdateContext);
-        break;
-
-      case 'permission':
-        restoredValues = rollingBackRollback
-          ? await redoPermissionChange(activity, preview.targetValues, effectiveSourceOperation, database)
-          : await rollbackPermissionChange(activity, database);
-        break;
-
-      case 'agent': {
-        const result = rollingBackRollback
-          ? await redoAgentConfigChange(activity, preview.targetValues, database, pageUpdateContext)
-          : await rollbackAgentConfigChange(activity, preview.currentValues, database, pageUpdateContext);
-        restoredValues = result.restoredValues;
-        pageMutationMeta = result.pageMutationMeta;
-        break;
-      }
-
-      case 'member':
-        restoredValues = rollingBackRollback
-          ? await redoMemberChange(activity, preview.targetValues, effectiveSourceOperation, database)
-          : await rollbackMemberChange(activity, database);
-        break;
-
-      case 'role':
-        restoredValues = rollingBackRollback
-          ? await redoRoleChange(activity, preview.targetValues, effectiveSourceOperation, database)
-          : await rollbackRoleChange(activity, database);
-        break;
-
-      case 'message':
-        restoredValues = rollingBackRollback
-          ? await redoMessageChange(activity, preview.targetValues, effectiveSourceOperation, database)
-          : await rollbackMessageChange(activity, database);
-        break;
-
-      case 'conversation':
-        // Conversation undo activities are logged for audit trail but cannot be rolled back
-        // because they represent the undo operation itself, not a data change
-        loggers.api.debug('[Rollback:Execute] Conversation undo cannot be rolled back', {
-          activityId: activity.id,
-          operation: activity.operation,
-        });
-        return {
-          success: false,
-          action: 'rollback',
-          status: 'failed',
-          message: 'Conversation undo operations cannot be rolled back. The affected messages remain soft-deleted and can be restored individually if needed.',
-          warnings,
-          changesApplied: preview.changes,
-        };
-
-      default:
-        loggers.api.debug('[Rollback:Execute] Unsupported resource type', {
-          resourceType: activity.resourceType,
-        });
-        return {
-          success: false,
-          action: 'rollback',
-          status: 'failed',
-          message: `Rollback not supported for resource type: ${activity.resourceType}`,
-          warnings,
-          changesApplied: preview.changes,
-        };
-    }
-
-    loggers.api.debug('[Rollback:Execute] Handler completed', {
-      resourceType: activity.resourceType,
-      restoredFieldsCount: Object.keys(restoredValues).length,
-    });
-
-    const resolvedContentSnapshot = await resolveActivityContentSnapshot(activity);
-
-    // Log the rollback activity with source snapshot for audit trail preservation
-    const logOptions: Parameters<typeof logRollbackActivity>[4] = {
-      restoredValues,
-      replacedValues: preview.currentValues ?? undefined,
-      contentSnapshot: resolvedContentSnapshot ?? undefined,
-      contentFormat: pageMutationMeta?.contentFormatAfter ?? activity.contentFormat ?? undefined,
-      // Source activity snapshot - survives retention policy deletion
-      // Use effectiveSourceOperation so "rollback of rollback" logs the original operation
-      rollbackSourceOperation: effectiveSourceOperation,
-      rollbackSourceTimestamp: rollingBackRollback
-        ? (activity.rollbackSourceTimestamp ?? activity.timestamp)
-        : activity.timestamp,
-      rollbackSourceTitle: activity.resourceTitle ?? undefined,
-      metadata: activity.metadata ? { sourceMetadata: activity.metadata } : undefined,
-      changeGroupId,
-      changeGroupType,
-      tx,
-    };
-
-    if (pageMutationMeta) {
-      logOptions.streamId = activity.pageId ?? activity.resourceId;
-      logOptions.streamSeq = pageMutationMeta.nextRevision;
-      logOptions.stateHashBefore = pageMutationMeta.stateHashBefore;
-      logOptions.stateHashAfter = pageMutationMeta.stateHashAfter;
-      logOptions.contentRef = pageMutationMeta.contentRefAfter ?? undefined;
-      logOptions.contentSize = pageMutationMeta.contentSizeAfter ?? undefined;
-    }
-
-    const deferredWorkflowTrigger = await logRollbackActivity(
-      userId,
-      activityId,
-      {
-        resourceType: activity.resourceType,
-        resourceId: activity.resourceId,
-        resourceTitle: activity.resourceTitle ?? undefined,
-        driveId: activity.driveId,
-        pageId: activity.pageId ?? undefined,
-      },
-      actorInfo,
-      logOptions
-    );
-
-    loggers.api.debug('[Rollback:Execute] Rollback completed successfully', {
-      activityId,
-      resourceType: activity.resourceType,
-      resourceId: activity.resourceId,
-    });
-
-    // Send notifications for newly mentioned users after all operations complete (fire-and-forget)
-    const mentionsResult = pageMutationMeta?.mentionsResult;
-    if (mentionsResult && mentionsResult.mentionedByUserId && mentionsResult.newlyMentionedUserIds.length > 0) {
-      for (const targetUserId of mentionsResult.newlyMentionedUserIds) {
-        createMentionNotification(targetUserId, mentionsResult.sourcePageId, mentionsResult.mentionedByUserId)
-          .catch((error: unknown) => {
-            loggers.api.error('Failed to send mention notification:', error as Error);
-          });
-      }
-    }
-
-    return {
-      success: true,
-      action: 'rollback',
-      status: 'success',
-      restoredValues,
-      message: 'Change undone',
-      warnings,
-      changesApplied: preview.changes,
-      deferredWorkflowTrigger,
-    };
-  } catch (error) {
-    loggers.api.error('[RollbackService] Error executing rollback', {
-      activityId,
-      userId,
-      error: error instanceof Error ? error.message : String(error),
-    });
-
-    return {
-      success: false,
-      action: 'rollback',
-      status: 'failed',
-      message: error instanceof Error ? error.message : 'Failed to execute rollback',
-      warnings,
-      changesApplied: preview.changes,
-    };
-  }
-}
-
-/**
- * Rollback a page change
- */
-async function rollbackPageChange(
-  activity: ActivityLogForRollback,
-  _currentValues: Record<string, unknown> | null,
-  database: typeof db,
-  pageUpdateContext: PageUpdateContext
-): Promise<PageChangeResult> {
-  loggers.api.debug('[Rollback:Execute:Page] Starting page rollback', {
-    pageId: activity.pageId,
-    operation: activity.operation,
-    updatedFields: activity.updatedFields,
-  });
-
-  if (!activity.pageId) {
-    throw new Error('Page ID not found in activity');
-  }
-
-  const resolvedContentSnapshot = await resolveActivityContentSnapshot(activity);
-  const plan = planPageRollback(activity, resolvedContentSnapshot);
-
-  // Handle create operation by trashing the page and orphaning its children.
-  if (plan.kind === 'trash-created') {
-    // Get the page's current state
-    const [page] = await database
-      .select({ parentId: pages.parentId, isTrashed: pages.isTrashed })
-      .from(pages)
-      .where(eq(pages.id, activity.pageId));
-
-    // If already trashed, skip - this is a no-op (handles AI undo of old activities)
-    if (page?.isTrashed) {
-      loggers.api.debug('[Rollback:Execute:Page] Page already trashed, skipping', {
-        pageId: activity.pageId,
-      });
-      return { restoredValues: { isTrashed: true }, pageMutationMeta: undefined };
-    }
-
-    loggers.api.debug('[Rollback:Execute:Page] Trashing created page', {
-      pageId: activity.pageId,
-    });
-
-    // Orphan any children to the grandparent (matches pageService.trashPage behavior)
-    // This prevents broken tree with children pointing to trashed parent
-    const childPages = await database
-      .select({ id: pages.id })
-      .from(pages)
-      .where(eq(pages.parentId, activity.pageId));
-
-    const nextParentId = page?.parentId ?? null;
-    for (const child of childPages) {
-      await applyPageUpdateWithRevision(database, child.id, {
-        parentId: nextParentId,
-        originalParentId: activity.pageId,
-      }, pageUpdateContext);
-    }
-
-    // Now trash the created page
-    const pageMutationMeta = await applyPageUpdateWithRevision(database, activity.pageId, {
-      isTrashed: true,
-      trashedAt: new Date(),
-    }, pageUpdateContext);
-
-    return { restoredValues: { isTrashed: true }, pageMutationMeta };
-  }
-
-  const { updateData, restoreOrphanedChildren } = plan;
-
-  loggers.api.debug('[Rollback:Execute:Page] Applying page update', {
-    pageId: activity.pageId,
-    fieldsToRestore: Object.keys(updateData),
-  });
-
-  // Update the page with revision/state hash
-  const pageMutationMeta = await applyPageUpdateWithRevision(database, activity.pageId, updateData, pageUpdateContext);
-
-  // If we're restoring a trashed page (isTrashed: false), also restore orphaned children
-  // When pages are trashed, children are orphaned to grandparent with originalParentId set
-  // Now that the parent is restored, re-parent those children back to their original parent
-  if (restoreOrphanedChildren) {
-    const restoredChildren = await database
-      .select({ id: pages.id })
-      .from(pages)
-      .where(eq(pages.originalParentId, activity.pageId));
-
-    for (const child of restoredChildren) {
-      await applyPageUpdateWithRevision(database, child.id, {
-        parentId: activity.pageId,
-        originalParentId: null,
-      }, pageUpdateContext);
-    }
-
-    if (restoredChildren.length > 0) {
-      loggers.api.debug('[Rollback:Execute:Page] Restored orphaned children', {
-        parentPageId: activity.pageId,
-        childrenRestored: restoredChildren.length,
-        childIds: restoredChildren.map(c => c.id),
-      });
-    }
-  }
-
-  return { restoredValues: updateData, pageMutationMeta };
-}
-
-/**
- * Rollback a drive change
- */
-async function rollbackDriveChange(
-  activity: ActivityLogForRollback,
-  _currentValues: Record<string, unknown> | null,
-  database: typeof db,
-  pageUpdateContext: PageUpdateContext
-): Promise<Record<string, unknown>> {
-  if (!activity.driveId) {
-    throw new Error('Drive ID not found in activity');
-  }
-
-  const plan = planDriveRollback(activity);
-
-  // Handle create operation by trashing the drive and all its pages
-  if (plan.kind === 'trash-created') {
-    loggers.api.debug('[Rollback:Execute:Drive] Trashing created drive and pages', {
-      driveId: activity.driveId,
-    });
-
-    const trashedAt = new Date();
-    const drivePages = await database
-      .select({ id: pages.id })
-      .from(pages)
-      .where(eq(pages.driveId, activity.driveId));
-
-    // First, trash all pages in the drive to prevent orphans
-    for (const page of drivePages) {
-      await applyPageUpdateWithRevision(database, page.id, {
-        isTrashed: true,
-        trashedAt,
-      }, pageUpdateContext);
-    }
-
-    // Then trash the drive itself
-    await database
-      .update(drives)
-      .set({
-        isTrashed: true,
-        trashedAt: new Date(),
-        updatedAt: new Date(),
-      })
-      .where(eq(drives.id, activity.driveId));
-
-    return { trashed: true, driveId: activity.driveId, pagesTrashed: true };
-  }
-
-  // Update the drive
-  await database
-    .update(drives)
-    .set({
-      ...plan.updateData,
-      updatedAt: new Date(),
-    })
-    .where(eq(drives.id, activity.driveId));
-
-  return plan.updateData;
-}
-
-/**
- * Rollback a permission change
- */
-async function rollbackPermissionChange(
-  activity: ActivityLogForRollback,
-  database: typeof db
-): Promise<Record<string, unknown>> {
-  const plan = planPermissionRollback(activity);
-
-  switch (plan.op) {
-    case 'delete': {
-      // A permission was granted - rollback by deleting it
-      await database
-        .delete(pagePermissions)
-        .where(and(eq(pagePermissions.pageId, plan.pageId), eq(pagePermissions.userId, plan.userId)));
-
-      loggers.api.info('[RollbackService] Deleted permission that was granted', {
-        pageId: plan.pageId,
-        userId: plan.userId,
-      });
-
-      return { deleted: true, pageId: plan.pageId, userId: plan.userId };
-    }
-
-    case 'insert': {
-      // A permission was revoked - rollback by re-creating it with previous values
-      await database.insert(pagePermissions).values(plan.values);
-
-      loggers.api.info('[RollbackService] Re-created revoked permission', {
-        pageId: plan.values.pageId,
-        userId: plan.values.userId,
-      });
-
-      return { ...plan.values };
-    }
-
-    case 'update': {
-      // A permission was updated - rollback by restoring previous values
-      await database
-        .update(pagePermissions)
-        .set(plan.set)
-        .where(and(eq(pagePermissions.pageId, plan.pageId), eq(pagePermissions.userId, plan.userId)));
-
-      loggers.api.info('[RollbackService] Restored previous permission values', {
-        pageId: plan.pageId,
-        userId: plan.userId,
-        restoredFields: Object.keys(plan.set),
-      });
-
-      return plan.set;
-    }
-  }
-}
-
-/**
- * Rollback an agent config change
- */
-async function rollbackAgentConfigChange(
-  activity: ActivityLogForRollback,
-  _currentValues: Record<string, unknown> | null,
-  database: typeof db,
-  pageUpdateContext: PageUpdateContext
-): Promise<PageChangeResult> {
-  // Agent configs are stored in pages table
-  const { updateData } = planAgentRollback(activity, AGENT_CONFIG_ROLLBACK_FIELDS);
-  const pageMutationMeta = await applyPageUpdateWithRevision(database, activity.pageId as string, updateData, pageUpdateContext);
-
-  return { restoredValues: updateData, pageMutationMeta };
-}
-
-/**
- * Rollback a member change
- */
-async function rollbackMemberChange(
-  activity: ActivityLogForRollback,
-  database: typeof db
-): Promise<Record<string, unknown>> {
-  const plan = planMemberRollback(activity, new Date());
-
-  switch (plan.op) {
-    case 'delete': {
-      // Member was added - rollback by removing them
-      await database
-        .delete(driveMembers)
-        .where(and(eq(driveMembers.driveId, plan.driveId), eq(driveMembers.userId, plan.userId)));
-
-      loggers.api.info('[RollbackService] Removed member that was added', {
-        driveId: plan.driveId,
-        userId: plan.userId,
-      });
-
-      return { deleted: true, driveId: plan.driveId, userId: plan.userId };
-    }
-
-    case 'insert': {
-      // Member was removed - rollback by re-adding them with previous values
-      await database.insert(driveMembers).values(plan.values);
-
-      loggers.api.info('[RollbackService] Re-added removed member', {
-        driveId: plan.values.driveId,
-        userId: plan.values.userId,
-        role: plan.values.role,
-      });
-
-      return { ...plan.values };
-    }
-
-    case 'update': {
-      // Role/customRole was changed - restore previous values
-      await database
-        .update(driveMembers)
-        .set(plan.set)
-        .where(and(eq(driveMembers.driveId, plan.driveId), eq(driveMembers.userId, plan.userId)));
-
-      loggers.api.info('[RollbackService] Restored previous member values', {
-        driveId: plan.driveId,
-        userId: plan.userId,
-        restoredFields: Object.keys(plan.set),
-      });
-
-      return plan.set;
-    }
-  }
-}
-
-/**
- * Rollback a role change
- */
-async function rollbackRoleChange(
-  activity: ActivityLogForRollback,
-  database: typeof db
-): Promise<Record<string, unknown>> {
-  const now = new Date();
-  const plan = planRoleRollback(activity, now);
-
-  switch (plan.op) {
-    case 'reorder': {
-      for (const [index, roleId] of plan.order.entries()) {
-        await database
-          .update(driveRoles)
-          .set({ position: index, updatedAt: now })
-          .where(eq(driveRoles.id, roleId));
-      }
-
-      loggers.api.info('[RollbackService] Restored previous role order', {
-        driveId: activity.driveId,
-        roleCount: plan.order.length,
-      });
-
-      return { order: plan.order };
-    }
-
-    case 'delete-role': {
-      // First, capture which members had this role for audit trail
-      const affectedMembers = await database
-        .select({ userId: driveMembers.userId })
-        .from(driveMembers)
-        .where(eq(driveMembers.customRoleId, plan.roleId));
-
-      // Delete the role (FK constraint will set customRoleId to null for affected members)
-      await database.delete(driveRoles).where(eq(driveRoles.id, plan.roleId));
-
-      loggers.api.info('[RollbackService] Deleted role that was created', {
-        driveId: activity.driveId,
-        roleId: plan.roleId,
-        affectedMemberCount: affectedMembers.length,
-      });
-
-      return {
-        deleted: true,
-        roleId: plan.roleId,
-        affectedMemberUserIds: affectedMembers.map(m => m.userId),
-      };
-    }
-
-    case 'insert-role': {
-      await database.insert(driveRoles).values(plan.values);
-
-      loggers.api.info('[RollbackService] Re-created deleted role', {
-        driveId: activity.driveId,
-        roleId: plan.values.id,
-        name: plan.values.name,
-      });
-
-      return { ...plan.values };
-    }
-
-    case 'update-role': {
-      await database
-        .update(driveRoles)
-        .set(plan.set)
-        .where(eq(driveRoles.id, plan.roleId));
-
-      loggers.api.info('[RollbackService] Restored previous role values', {
-        driveId: activity.driveId,
-        roleId: plan.roleId,
-        restoredFields: Object.keys(plan.set),
-      });
-
-      return plan.set;
-    }
-  }
-}
-
-/**
- * Rollback a message change (edit or delete)
- */
-async function rollbackMessageChange(
-  activity: ActivityLogForRollback,
-  database: typeof db
-): Promise<Record<string, unknown>> {
-  const messageId = activity.resourceId;
-  const metadata = activity.metadata as Record<string, unknown> | null;
-  const conversationType = metadata?.conversationType as string | undefined;
-
-  // Determine which table to update based on conversationType
-  // Channel messages use the channelMessages table; global uses messages; page uses chatMessages.
-  // channelMessages has no editedAt column, so a channel edit restores content only —
-  // planMessageRollback encodes that in `set` while keeping a uniform return value.
-  const { table, isChannel, label: tableLabel } = pickConversationTable({
-    conversationType,
-    hasPageId: !!activity.pageId,
-  });
-
-  const plan = planMessageRollback(activity, isChannel);
-
-  await database
-    .update(table)
-    .set(plan.set)
-    .where(eq(table.id, messageId));
-
-  loggers.api.info(`[RollbackService] Applied message rollback (${tableLabel})`, {
-    messageId,
-    operation: activity.operation,
-    pageId: activity.pageId,
-  });
-
-  return plan.returnValue;
-}
-
-/**
- * Redo a page change (undo a rollback)
- */
-async function redoPageChange(
-  activity: ActivityLogForRollback,
-  targetValues: Record<string, unknown> | null,
-  sourceOperation: ActivityOperation,
-  database: typeof db,
-  pageUpdateContext: PageUpdateContext
-): Promise<PageChangeResult> {
-  if (!activity.pageId) {
-    throw new Error('Page ID not found in activity');
-  }
-
-  const updateData = planPageRedo(targetValues, sourceOperation);
-
-  if (updateData.isTrashed === true) {
-    const [page] = await database
-      .select({ parentId: pages.parentId })
-      .from(pages)
-      .where(eq(pages.id, activity.pageId));
-
-    const childPages = await database
-      .select({ id: pages.id })
-      .from(pages)
-      .where(eq(pages.parentId, activity.pageId));
-
-    const nextParentId = page?.parentId ?? null;
-    for (const child of childPages) {
-      await applyPageUpdateWithRevision(database, child.id, {
-        parentId: nextParentId,
-        originalParentId: activity.pageId,
-      }, pageUpdateContext);
-    }
-
-    updateData.trashedAt = new Date();
-  }
-
-  if (updateData.isTrashed === false) {
-    updateData.trashedAt = null;
-  }
-
-  const pageMutationMeta = await applyPageUpdateWithRevision(database, activity.pageId, updateData, pageUpdateContext);
-
-  if (updateData.isTrashed === false) {
-    const restoredChildren = await database
-      .select({ id: pages.id })
-      .from(pages)
-      .where(eq(pages.originalParentId, activity.pageId));
-
-    for (const child of restoredChildren) {
-      await applyPageUpdateWithRevision(database, child.id, {
-        parentId: activity.pageId,
-        originalParentId: null,
-      }, pageUpdateContext);
-    }
-  }
-
-  return { restoredValues: updateData, pageMutationMeta };
-}
-
-/**
- * Redo a drive change (undo a rollback)
- */
-async function redoDriveChange(
-  activity: ActivityLogForRollback,
-  targetValues: Record<string, unknown> | null,
-  sourceOperation: ActivityOperation,
-  database: typeof db,
-  pageUpdateContext: PageUpdateContext
-): Promise<Record<string, unknown>> {
-  if (!activity.driveId) {
-    throw new Error('Drive ID not found in activity');
-  }
-
-  const updateData = planDriveRedo(targetValues, sourceOperation);
-
-  if (updateData.isTrashed === true) {
-    const trashedAt = new Date();
-    const drivePages = await database
-      .select({ id: pages.id })
-      .from(pages)
-      .where(eq(pages.driveId, activity.driveId));
-
-    for (const page of drivePages) {
-      await applyPageUpdateWithRevision(database, page.id, {
-        isTrashed: true,
-        trashedAt,
-      }, pageUpdateContext);
-    }
-    updateData.trashedAt = trashedAt;
-  }
-
-  if (updateData.isTrashed === false) {
-    const drivePages = await database
-      .select({ id: pages.id })
-      .from(pages)
-      .where(eq(pages.driveId, activity.driveId));
-
-    for (const page of drivePages) {
-      await applyPageUpdateWithRevision(database, page.id, {
-        isTrashed: false,
-        trashedAt: null,
-      }, pageUpdateContext);
-    }
-    updateData.trashedAt = null;
-  }
-
-  await database
-    .update(drives)
-    .set({
-      ...updateData,
-      updatedAt: new Date(),
-    })
-    .where(eq(drives.id, activity.driveId));
-
-  return updateData;
-}
-
-/**
- * Redo a permission change (undo a rollback)
- */
-async function redoPermissionChange(
-  activity: ActivityLogForRollback,
-  targetValues: Record<string, unknown> | null,
-  sourceOperation: ActivityOperation,
-  database: typeof db
-): Promise<Record<string, unknown>> {
-  const plan = planPermissionRedo(activity, targetValues, sourceOperation);
-
-  switch (plan.op) {
-    case 'upsert': {
-      await database
-        .insert(pagePermissions)
-        .values(plan.values)
-        .onConflictDoUpdate({
-          target: [pagePermissions.pageId, pagePermissions.userId],
-          set: plan.values,
-        });
-
-      return { ...plan.values };
-    }
-
-    case 'update': {
-      await database
-        .update(pagePermissions)
-        .set(plan.set)
-        .where(and(eq(pagePermissions.pageId, plan.pageId), eq(pagePermissions.userId, plan.userId)));
-
-      return plan.set;
-    }
-
-    case 'delete': {
-      await database
-        .delete(pagePermissions)
-        .where(and(eq(pagePermissions.pageId, plan.pageId), eq(pagePermissions.userId, plan.userId)));
-
-      return { deleted: true, pageId: plan.pageId, userId: plan.userId };
-    }
-  }
-}
-
-/**
- * Redo a member change (undo a rollback)
- */
-async function redoMemberChange(
-  activity: ActivityLogForRollback,
-  targetValues: Record<string, unknown> | null,
-  sourceOperation: ActivityOperation,
-  database: typeof db
-): Promise<Record<string, unknown>> {
-  const plan = planMemberRedo(activity, targetValues, sourceOperation, new Date());
-
-  switch (plan.op) {
-    case 'upsert': {
-      await database
-        .insert(driveMembers)
-        .values(plan.values)
-        .onConflictDoUpdate({
-          target: [driveMembers.driveId, driveMembers.userId],
-          set: plan.values,
-        });
-
-      return { ...plan.values };
-    }
-
-    case 'delete': {
-      await database
-        .delete(driveMembers)
-        .where(and(eq(driveMembers.driveId, plan.driveId), eq(driveMembers.userId, plan.userId)));
-
-      return { deleted: true, driveId: plan.driveId, userId: plan.userId };
-    }
-
-    case 'update': {
-      await database
-        .update(driveMembers)
-        .set(plan.set)
-        .where(and(eq(driveMembers.driveId, plan.driveId), eq(driveMembers.userId, plan.userId)));
-
-      return plan.set;
-    }
-  }
-}
-
-/**
- * Redo a role change (undo a rollback)
- */
-async function redoRoleChange(
-  activity: ActivityLogForRollback,
-  targetValues: Record<string, unknown> | null,
-  sourceOperation: ActivityOperation,
-  database: typeof db
-): Promise<Record<string, unknown>> {
-  const now = new Date();
-  const plan = planRoleRedo(activity, targetValues, sourceOperation, now);
-
-  switch (plan.op) {
-    case 'reorder': {
-      for (const [index, targetRoleId] of plan.order.entries()) {
-        await database
-          .update(driveRoles)
-          .set({ position: index, updatedAt: now })
-          .where(eq(driveRoles.id, targetRoleId));
-      }
-
-      return { order: plan.order };
-    }
-
-    case 'insert-role': {
-      await database.insert(driveRoles).values(plan.values);
-      return { ...plan.values };
-    }
-
-    case 'delete-role': {
-      const affectedMembers = await database
-        .select({ userId: driveMembers.userId })
-        .from(driveMembers)
-        .where(eq(driveMembers.customRoleId, plan.roleId));
-
-      await database.delete(driveRoles).where(eq(driveRoles.id, plan.roleId));
-
-      return {
-        deleted: true,
-        roleId: plan.roleId,
-        affectedMemberUserIds: affectedMembers.map(member => member.userId),
-      };
-    }
-
-    case 'update-role': {
-      await database
-        .update(driveRoles)
-        .set(plan.set)
-        .where(eq(driveRoles.id, plan.roleId));
-
-      return plan.set;
-    }
-  }
-}
-
-/**
- * Redo agent config change (undo a rollback)
- */
-async function redoAgentConfigChange(
-  activity: ActivityLogForRollback,
-  targetValues: Record<string, unknown> | null,
-  database: typeof db,
-  pageUpdateContext: PageUpdateContext
-): Promise<PageChangeResult> {
-  const { updateData } = planAgentRedo(activity, targetValues, AGENT_CONFIG_ROLLBACK_FIELDS);
-  const pageMutationMeta = await applyPageUpdateWithRevision(database, activity.pageId as string, updateData, pageUpdateContext);
-
-  return { restoredValues: updateData, pageMutationMeta };
-}
-
-/**
- * Redo a message change (undo a rollback)
- */
-async function redoMessageChange(
-  activity: ActivityLogForRollback,
-  targetValues: Record<string, unknown> | null,
-  sourceOperation: ActivityOperation,
-  database: typeof db
-): Promise<Record<string, unknown>> {
-  const metadata = activity.metadata as Record<string, unknown> | null;
-  const conversationType = metadata?.conversationType as string | undefined;
-  const { table, isChannel } = pickConversationTable({ conversationType, hasPageId: !!activity.pageId });
-
-  const updateData = planMessageRedo(targetValues, sourceOperation, isChannel, new Date());
-
-  await database
-    .update(table)
-    .set(updateData)
-    .where(eq(table.id, activity.resourceId));
-
-  return updateData;
+  return shellExecuteRollback(deps, activityId, userId, context, options)
 }
 
 /**
@@ -2060,56 +77,7 @@ export async function getPageVersionHistory(
   userId: string,
   options: VersionHistoryOptions = {}
 ): Promise<{ activities: ActivityLogForRollback[]; total: number }> {
-  const { limit = 50, offset = 0, startDate, endDate, actorId, operation, includeAiOnly } = options;
-
-  loggers.api.debug('[History:Fetch] Fetching page version history', {
-    pageId,
-    userId,
-    limit,
-    offset,
-    hasFilters: !!(startDate || endDate || actorId || operation || includeAiOnly),
-  });
-
-  try {
-    const conditions = buildHistoryConditions(eq(activityLogs.pageId, pageId), {
-      startDate,
-      endDate,
-      actorId,
-      operation,
-      includeAiOnly,
-    });
-
-    const [activities, countResult] = await Promise.all([
-      db
-        .select()
-        .from(activityLogs)
-        .where(and(...conditions))
-        .orderBy(desc(activityLogs.timestamp))
-        .limit(limit)
-        .offset(offset),
-      db
-        .select({ value: count() })
-        .from(activityLogs)
-        .where(and(...conditions)),
-    ]);
-
-    loggers.api.debug('[History:Fetch] Page history query complete', {
-      pageId,
-      activitiesCount: activities.length,
-      total: countResult[0]?.value ?? 0,
-    });
-
-    return {
-      activities: activities.map(mapActivityRow),
-      total: countResult[0]?.value ?? 0,
-    };
-  } catch (error) {
-    loggers.api.error('[RollbackService] Error fetching page version history', {
-      pageId,
-      error: error instanceof Error ? error.message : String(error),
-    });
-    return { activities: [], total: 0 };
-  }
+  return repoGetPageVersionHistory(deps, pageId, userId, options)
 }
 
 /**
@@ -2120,89 +88,12 @@ export async function getDriveVersionHistory(
   userId: string,
   options: VersionHistoryOptions = {}
 ): Promise<{ activities: ActivityLogForRollback[]; total: number }> {
-  const { limit = 50, offset = 0, startDate, endDate, actorId, operation, resourceType } = options;
-
-  loggers.api.debug('[History:Fetch] Fetching drive version history', {
-    driveId,
-    userId,
-    limit,
-    offset,
-    hasFilters: !!(startDate || endDate || actorId || operation || resourceType),
-  });
-
-  try {
-    const conditions = buildHistoryConditions(eq(activityLogs.driveId, driveId), {
-      startDate,
-      endDate,
-      actorId,
-      operation,
-      resourceType,
-    });
-
-    const [activities, countResult] = await Promise.all([
-      db
-        .select()
-        .from(activityLogs)
-        .where(and(...conditions))
-        .orderBy(desc(activityLogs.timestamp))
-        .limit(limit)
-        .offset(offset),
-      db
-        .select({ value: count() })
-        .from(activityLogs)
-        .where(and(...conditions)),
-    ]);
-
-    loggers.api.debug('[History:Fetch] Drive history query complete', {
-      driveId,
-      activitiesCount: activities.length,
-      total: countResult[0]?.value ?? 0,
-    });
-
-    return {
-      activities: activities.map(mapActivityRow),
-      total: countResult[0]?.value ?? 0,
-    };
-  } catch (error) {
-    loggers.api.error('[RollbackService] Error fetching drive version history', {
-      driveId,
-      error: error instanceof Error ? error.message : String(error),
-    });
-    return { activities: [], total: 0 };
-  }
+  return repoGetDriveVersionHistory(deps, driveId, userId, options)
 }
 
 /**
  * Get user's retention limit based on subscription tier
  */
 export async function getUserRetentionDays(userId: string): Promise<number> {
-  // Default retention days by tier (ordered: free < pro < founder < business)
-  const defaultRetention: Record<string, number> = {
-    free: 7,
-    pro: 30,
-    founder: 90,
-    business: -1, // unlimited
-  };
-
-  try {
-    // Get user's subscription tier
-    const user = await db
-      .select({ subscriptionTier: users.subscriptionTier })
-      .from(users)
-      .where(eq(users.id, userId))
-      .limit(1);
-
-    if (user.length === 0) {
-      return defaultRetention.free;
-    }
-
-    const tier = user[0].subscriptionTier || 'free';
-    return defaultRetention[tier] || defaultRetention.free;
-  } catch (error) {
-    loggers.api.error('[RollbackService] Error getting user retention days', {
-      userId,
-      error: error instanceof Error ? error.message : String(error),
-    });
-    return defaultRetention.free;
-  }
+  return repoGetUserRetentionDays(deps, userId)
 }
